refactor(Filters): unify font and color styles across all filters and consolidate code to avoid duplication
